requirements.txt
py-clob-client>=0.16.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
